+++ models/classification/_base_/efficientnet_b0_omz.py
@@ -12,7 +12,7 @@ model = dict(
     neck=dict(
         type='GlobalAveragePooling'),
     head=dict(
-        type='CustomLinearClsHead',
+        type='LinearClsHead',
         num_classes=1000,
         in_channels=320,
         loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
