+++ models/classification/_base_/mobilenet_v2_omz.py
@@ -12,7 +12,7 @@ model = dict(
     neck=dict(
         type='GlobalAveragePooling'),
     head=dict(
-        type='CustomLinearClsHead',
+        type='LinearClsHead',
         num_classes=1000,
         in_channels=1280,
         loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
