+++ samples/cfgs/models/heads/linear_cls.yaml
@@ -1,6 +1,6 @@
 model:
     head:
-        type: CustomLinearClsHead
+        type: LinearClsHead
         num_classes: -1
         in_channels: -1
         topk: !!python/tuple [1, 5]
